rolled back some OpenAI and pandas parsers
